Implement review changes
