+++ core/templates/core/partials/create_project.html
@@ -1,8 +1,4 @@
 {% load i18n %}
-{% comment %}
-Changed implementation to using html native modal. This automatically disables the whole page
-and allows easy interaction. For now the Modal does not delete input when the modal is canceled
-{% endcomment %}
 <dialog id="{{id}}"
         closedby='any'
         class="
