+++ core/views.py
@@ -18,10 +18,10 @@ from django.urls import reverse
 from django.utils.translation import gettext as _
 from django.views.generic import TemplateView
 
+from ports.authorization import has_authorization
 from ports.forms import CreateProjectForm
 from ports.models import Project
 from ports.models import Scenario
-from ports.models import has_authorization
 from ports.util import get_template_scenarios
 from ports.util import get_user_projects
 from ports.util import prefetch_projects_users
@@ -30,11 +30,8 @@ from .forms import AuthForm
 from .forms import SignUpForm
 
 
+@login_required()
 def projects_view(request):
-    if not request.user.is_authenticated:
-        # login may further redirect to the
-        # LOGIN_REDIRECT_URL
-        return redirect(reverse("core:login"))
     context = {}
     template_scenarios = get_template_scenarios(request.user)
     projects = get_user_projects(request.user)
@@ -45,6 +42,7 @@ def projects_view(request):
 
 
 def scenario_results(request, scenario_internal_id):
+    # TODO: guard results page against unwarranted access
     context = {}
     scenario = get_object_or_404(Scenario, internal_id=scenario_internal_id)
     context["scenario"] = scenario
@@ -53,11 +51,8 @@ def scenario_results(request, scenario_internal_id):
     return render(request, template_name="core/ergebnisse.html", context=context)
 
 
+@login_required()
 def project_overview_view(request, project_internal_id):
-    if not request.user.is_authenticated:
-        # login may further redirect to the
-        # LOGIN_REDIRECT_URL
-        return redirect(reverse("core:login"))
     project = get_object_or_404(Project, internal_id=project_internal_id)
 
     if not has_authorization(project, request.user, "view"):
@@ -135,7 +130,7 @@ def signup(request):
     raise Http404()
 
 
-@login_required(login_url="/login/")
+@login_required()
 def changePassword(request):
     if request.method == "POST":
         form = PasswordChangeForm(request.user, request.POST)
@@ -152,7 +147,7 @@ def changePassword(request):
     return render(request, "core/registration/password_change.html", {"form": form})
 
 
-@login_required(login_url="/login/")
+@login_required()
 def test_email(request):
     if request.user.is_staff:
         mail.send_mail(
