+++ enetra/settings.py
@@ -213,6 +213,7 @@ EMAIL_HOST_PASSWORD = env("EMAIL_HOST_PASSWORD", default=None)
 EMAIL_USE_TLS = True
 DEFAULT_FROM_EMAIL = EMAIL_HOST_USER
 
+LOGIN_URL = "core:login"
 LOGIN_REDIRECT_URL = "/"  # redirect to landing page after login
 LOGOUT_REDIRECT_URL = "/"  # redirect to landing page after logout as well
