+++ ports/models.py
@@ -4,7 +4,6 @@ from collections.abc import Iterable
 from contextlib import contextmanager
 from pathlib import Path
 from typing import ClassVar
-from typing import Literal
 
 from django.conf import settings
 from django.contrib.auth.models import User
@@ -14,9 +13,6 @@ from django.contrib.gis.db import models
 from django.core.files.uploadedfile import InMemoryUploadedFile
 from django.core.validators import MaxValueValidator
 from django.core.validators import MinValueValidator
-from django.db.models import ForeignKey
-from django.db.models import ManyToManyField
-from django.db.models import QuerySet
 from django.db.models.functions import Now
 from django.db.models.signals import post_delete
 from django.db.models.signals import post_save
@@ -26,7 +22,6 @@ from django.dispatch import receiver
 from django.forms import ModelForm
 from django.utils.safestring import mark_safe
 from django.utils.translation import gettext_lazy as _
-from guardian.shortcuts import get_objects_for_user
 from guardian.utils import get_group_obj_perms_model
 
 logger = logging.getLogger(__name__)
@@ -45,6 +40,7 @@ class Project(models.Model):
     updated_at = models.DateTimeField(auto_now=True)
 
     # Related name + tells django not to create a reverse relation for user, e.g. user.scenario_set
+    # To find all projects a user has access to use util.get_user_projects, which takes group permissions into account
     manager = models.ForeignKey(
         User, on_delete=models.SET_NULL, default=None, null=True, related_name="+"
     )
@@ -82,8 +78,7 @@ class Project(models.Model):
     @atomic()
     def safe_delete(self):
         """Safely delete the project by safely deleting all scenarios referencing it.
-        This is needed because of the nature of DeletedItems which are created during
-        deletion
+        This is needed because of DeletedItems which are created during deletion.
         """
         scenarios = Scenario.objects.filter(project=self)
         for s in scenarios:
@@ -91,11 +86,13 @@ class Project(models.Model):
         self.delete()
 
 
-# Each set of scenario items is bundled via its scenario. The scenario has a simple BigInteger Id
 class Scenario(models.Model):
-    # Project which bundles Scenarios
-    # on_delete is null since DeletedItems dont allow cascading delete on Scenario
-    # instead safe_delete has to be used on the scenarios or project
+    # Instance which bundles ScenarioItems
+    # Bundled with other scenarios in a common _Project_.
+
+    # FK project is set to NULL when project is deleted,
+    # since DeletedItems don't allow cascading delete.
+    # Use project.safe_delete instead.
     project = models.ForeignKey(Project, on_delete=models.SET_NULL, default=None, null=True)
     id = models.BigAutoField(primary_key=True, blank=True)
     internal_id = models.UUIDField(
@@ -150,7 +147,9 @@ class Scenario(models.Model):
             self.delete()
 
     def prepare_deepcopy(self):
-        """Adjust the instance so it can be deepcopied"""
+        """Adjust the instance so it can be deepcopied
+        Since the internal_id of each scenario has to be unique, its overwritten with a new one.
+        """
         self.internal_id = uuid.uuid4()
 
     def changed_event(self):
@@ -1002,100 +1001,6 @@ def auto_delete_file_on_delete(sender, instance, **kwargs):
         path = Path(instance.file.path)
         if path.exists():
             path.unlink()
-
-
-def get_related_model_values(
-    instances: QuerySet[ScenarioItem], model: type[ScenarioItem], field_name: str
-):
-    """Find the values of field_name on all related instances of type model for the given queryset
-
-    Used to find all area_internal_ids for a given QuerySet of ScenarioItems, so the permission
-    can be checked on those areas
-    """
-    all_fields = instances.model._meta.get_fields()
-    fk_fields = [
-        f for f in all_fields if isinstance(f, ForeignKey) and issubclass(f.related_model, model)
-    ]
-    m2m_fields = [
-        f
-        for f in all_fields
-        if isinstance(f, ManyToManyField) and issubclass(f.related_model, model)
-    ]
-    if not fk_fields and not m2m_fields:
-        return None
-
-    if fk_fields:
-        instances = instances.select_related(*[f.name for f in fk_fields])
-    if m2m_fields:
-        instances = instances.prefetch_related(*[f.name for f in m2m_fields])
-
-    values = []
-    for instance in instances:
-        for field in fk_fields:
-            related_instance = getattr(instance, field.name)
-            if related_instance is not None:
-                values.append(getattr(related_instance, field_name))
-        for field in m2m_fields:
-            values.extend(getattr(instance, field.name).values_list(field_name, flat=True))
-    return values
-
-
-def has_area_authorization_from_uuids(
-    uuids: list[uuid.UUID | str],
-    user: User,
-    crud: Literal["view", "create", "details", "change", "delete"],
-    scenario: Scenario,
-    model: type[ScenarioItem],
-):
-    if model is Area:
-        area_uuids = uuids
-    else:
-        instances = model.objects.filter(scenario=scenario, internal_id__in=uuids)
-        area_uuids = get_related_model_values(instances, Area, "internal_id")
-    areas = Area.objects.filter(scenario=scenario, internal_id__in=area_uuids).exclude(manager=user)
-    non_managed_areas_internal_ids = areas.values_list("internal_id", flat=True)
-    allowed_internal_ids = get_objects_for_user(user, crud, areas).values_list(
-        "internal_id", flat=True
-    )
-    missing_permissions = len(set(non_managed_areas_internal_ids).difference(allowed_internal_ids))
-    return missing_permissions == 0
-
-
-def has_authorization(
-    item: Project | ScenarioItem,
-    user: User,
-    crud: Literal["view", "create", "details", "change", "delete"],
-) -> bool:
-    """
-    View: See an item with limited set of attributes
-    details: See an item with all its attributes
-    Change: Change an item
-    Delete: Delete an item
-    Create: Create an item
-    """
-    if not user.is_authenticated:
-        return False
-    if user.is_superuser:
-        return True
-    if user == item.manager:
-        return True
-
-    # For now short the authorization for ScenarioItems, since their authorization is
-    # based only on areas for now
-    if isinstance(item, ScenarioItem):
-        return has_area_authorization_from_uuids(
-            [item.internal_id], user, crud, item.scenario, item.model()
-        )
-
-    has_perm = False
-    match crud:
-        case "delete" | "details":
-            has_perm = user.has_perm("details", item)
-        case "view":
-            has_perm = user.has_perm("view", item)
-        case _:
-            print("no matching crud found for " + crud)
-    return has_perm
 
 
 # Custom Exceptions
